Refactored builders to use AM6Client vs io.proxy
First, this is a cruddy refactor job - there's so much duplication in
here it's sick.  Second, while I like the io.proxy concept, the way it's
handled through the bus service, over the years, was just too
cumbersome.  While I liked the idea of the services shifting internally
to a common request mechanism and object model, regardless of the
service, it was just too much of a pain to maintain and actually use.
On the flip side, taking this out stripped a lot of the dynacism of the
path expressions, particularly when compiling the components into a DWAC
publication.  Oh well. There's a bunch of legacy/dead code in the core
framework that can get pruned, but, frankly, since few if any people
actually bother at the moment then I'll just leave it as is - if I
decide to really refactor it, I'll probably just rewrite the whole thing
again to use more modern design and development practices as this one
uses a pretty old system that's nearly eighteen years old.