Bug fix in handling empty string and error conditions for JSON operations
The expected behavior in this library is the error as left in the cache while the caller only receives the expected value or a null.  The expected value on an asynchronous call is an internal object with a pointer to the json object, and if that internal object is null an error occurred or invalid request was made.  Since that is logged here and the result cached, the caller should ignore it and treat is as a simple check for the object existing and the property being set.